+++ compute/zsyr2k.c
@@ -169,7 +169,7 @@ int PLASMA_zsyr2k(PLASMA_enum uplo, PLASMA_enum trans,
     }
 
     // quick return
-    if (n == 0 || ((alpha == zzero || k == 0.0) && beta == (double)1.0))
+    if (n == 0 || ((alpha == zzero || k == 0.0) && beta == 1.0))
         return PLASMA_SUCCESS;
 
     // Tune
@@ -420,7 +420,7 @@ void PLASMA_zsyr2k_Tile_Async(PLASMA_enum uplo, PLASMA_enum trans,
     }
 
     // quick return
-    if (C->m == 0 || ((alpha == zzero || An == 0) && beta == (double)1.0))
+    if (C->m == 0 || ((alpha == zzero || An == 0) && beta == 1.0))
         return;
 
     // Call the parallel function.
